fix boost::posix_time::seconds on ubuntu disco with boost 1.67

+++ source/type/pb_converter.cpp
@@ -1907,7 +1907,7 @@ void PbCreator::finalize_section(pbnavitia::Section* section,
     section->set_length(total_length);
 
     section->set_begin_date_time(navitia::to_posix_timestamp(departure));
-    section->set_end_date_time(navitia::to_posix_timestamp(departure + bt::seconds(total_duration)));
+    section->set_end_date_time(navitia::to_posix_timestamp(departure + bt::seconds(long(total_duration))));
 
     // add the destination as a placemark
     pbnavitia::PtObject* dest_place = section->mutable_destination();
